WIP: attempt at making it possible to install custom server parts
This doesn't work; the install process needs to run in the "screeps" layer
of the Dockerfile, because driver might needs python2 & gcc and stuff.
But at that point of the image building we don't have config.yml, so
it's not like a script is gonna be able to perform the fake install.

The WIP is that this needs a copy of config.yml inside ./server so that
it can be COPY-ed into the builder image. But even though that was
enough to make the install script run, it wasn't sufficient to actually
properly install stuff? Not sure where my changes went there…